# server/swagger/swagger.yml
basePath: /api/v2
consumes:
    - application/json
definitions:
...
    Board:
        description: Board groups a set of blocks and its layout
        x-go-package: github.com/nikethai/focalboard/server/model
...
    BoardMetadata:
        description: BoardMetadata contains metadata for a Board
        x-go-package: github.com/nikethai/focalboard/server/model
    BoardPatch:
        description: BoardPatch is a patch for modify boards
        x-go-package: github.com/nikethai/focalboard/server/model
    BoardsAndBlocks:
        description: |-
            BoardsAndBlocks is used to operate over boards and blocks at the
            same time
        x-go-package: github.com/nikethai/focalboard/server/model
...
    ErrorResponse:
        description: ErrorResponse is an error response
        x-go-package: github.com/nikethai/focalboard/server/model
...
host: localhost
info:
    contact:
        email: api@focalboard.com
        name: Focalboard
        url: https://www.focalboard.com
    description: Focalboard Server
    license:
        name: Custom
        url: https://github.com/nikethai/focalboard/blob/main/LICENSE.txt
    title: Focalboard Server
    version: 2.0.0
paths:
...
    /boards:
        post:
            description: Creates a new board
            operationId: createBoard
            parameters:
                - description: the board to create
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/Board'
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/Board'
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
    /boards/{boardID}:
        delete:
            description: Removes a board
            operationId: deleteBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        get:
            description: Returns a board
            operationId: getBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/Board'
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        patch:
            description: Partially updates a board
            operationId: patchBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: board patch to apply
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/BoardPatch'
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/Board'
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
    /boards/{boardID}/duplicate:
        post:
            description: Returns the new created board and all the blocks
            operationId: duplicateBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/BoardsAndBlocks'
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
    /boards/{boardID}/metadata:
        get:
            description: Returns a board's metadata
            operationId: getBoardMetadata
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/BoardMetadata'
                "404":
                    description: board not found
                "501":
                    description: required license not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
    /boards/{boardID}/undelete:
        post:
            description: Undeletes a board
            operationId: undeleteBoard
            parameters:
                - description: ID of board to undelete
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
    /teams/{teamID}/boards:
        get:
            description: Returns team boards
            operationId: getBoards
            parameters:
                - description: Team ID
                  in: path
                  name: teamID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        items:
                            $ref: '#/definitions/Board'
                        type: array
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
    /teams/{teamID}/categories/{categoryID}/boards/{boardID}:
        post:
            description: Set the category of a board
            operationId: updateCategoryBoard
            parameters:
                - description: Team ID
                  in: path
                  name: teamID
                  required: true
                  type: string
                - description: Category ID
                  in: path
                  name: categoryID
                  required: true
                  type: string
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
produces:
    - application/json
schemes:
    - http
    - https
securityDefinitions:
    BearerAuth:
        description: 'Pass session token using Bearer authentication, e.g. set header "Authorization: Bearer <session token>"'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"